Worked on calendar
Added a time slider above the calendar to control what the next added availability period will be. Made the copy button on a time period copy the time of an existing period onto the time slider above the calendar.
