+++ OpenShift/Availability.aspx
@@ -8,6 +8,7 @@
 </asp:Content>
 
 <asp:Content ID="Content1" ContentPlaceHolderID="MainContent" runat="Server">
+    <h1 class="display-4 text-center p-4">Availability</h1>
     <div id="availability-calendar"></div>
     <script type="module">
         import { AvailabilityCalendar } from "./Scripts/calendar.js";
@@ -21,8 +22,7 @@
         let workingHoursEnd = PLACEHOLDER;
 
         let container = document.getElementById( "availability-calendar" );
-        let calendar = new AvailabilityCalendar( "17:00", "24:00", 15 );
+        let calendar = new AvailabilityCalendar( "17:00", "24:00", 5 );
         calendar.appendTo( container );
     </script>
 </asp:Content>
-
